Moved HTTP request body conversion to a separate class, with its first test.
